backend/requirements.txt
flask==2.3.2
opencv-python-headless==4.7.0.72
deepface==0.0.79
numpy==1.24.3
gunicorn==20.1.0
python-dotenv==1.0.0
flask-cors==3.0.10